Move removeDeletedNode to where it's used
- Fixes circular dependency
